requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
python-jose
python-dotenv
python-multipart
reportlab
psycopg2-binary
cloudinary